# reduce data gather unless need for wildcard matching