@SEJeff 2017-12-14
FWIW: I confirmed this patch and it works really well. The original behavior was super unpleasant.

cc: @terminalmage who's done lots of work around fixing this module awhile ago.