@rallytime 2017-12-14
Hi @junster1 - Thanks for submitting this. The `2017.7.2` branch is closed for changes. Can you move this to `2017.7` instead? That way it can be included in `2017.7.3`.