@junster1 2017-12-14
Hmm.. seems that this is already fixed in that branch, just a slightly different way, but timing was the same.

commit b1e64b11fbcaf9dedb6d4cb6ee7a06801cbba877
Author: Michal Kurtak <michal.kurtak@gmail.com>
Date:   Thu Sep 21 21:51:22 2017 +0200

    yumpkg.py: install calls list_repo_pkgs only if wildcard in pkg name is used
    
    Fixes #43396